eventhub: consolidating onto API Version 2021-11-01
